+++ func/genBatchPDF.php
@@ -16,6 +16,12 @@ function genBatchPDF($fid, $batchID, $bottle, $new_conc, $mg, $ver, $uploads_pat
 			$this->SetFont('Arial','I',8);
 			$this->Cell(250,5,$batchID,0,0,'C');
 			$this->Ln(20);
+			//Barcode
+			$this->SetY(12);
+			$this->SetFont('Arial','I',8);
+			$this->Code39(80,20,$batchID,1,10);
+			$this->Ln(30);
+
 		}
 				
 		function Footer() {
@@ -24,7 +30,85 @@ function genBatchPDF($fid, $batchID, $bottle, $new_conc, $mg, $ver, $uploads_pat
 			$this->SetFont('Arial','I',8);
 			$this->Cell(0,10,'Page '.$this->PageNo().'/{nb} - Perfumers Vault Pro '.$ver,0,0,'C');
 		}
-	}		
+
+		function Code39($xpos, $ypos, $code, $baseline=0.5, $height=5){
+
+			$wide = $baseline;
+			$narrow = $baseline / 3 ; 
+			$gap = $narrow;
+		
+			$barChar['0'] = 'nnnwwnwnn';
+			$barChar['1'] = 'wnnwnnnnw';
+			$barChar['2'] = 'nnwwnnnnw';
+			$barChar['3'] = 'wnwwnnnnn';
+			$barChar['4'] = 'nnnwwnnnw';
+			$barChar['5'] = 'wnnwwnnnn';
+			$barChar['6'] = 'nnwwwnnnn';
+			$barChar['7'] = 'nnnwnnwnw';
+			$barChar['8'] = 'wnnwnnwnn';
+			$barChar['9'] = 'nnwwnnwnn';
+			$barChar['A'] = 'wnnnnwnnw';
+			$barChar['B'] = 'nnwnnwnnw';
+			$barChar['C'] = 'wnwnnwnnn';
+			$barChar['D'] = 'nnnnwwnnw';
+			$barChar['E'] = 'wnnnwwnnn';
+			$barChar['F'] = 'nnwnwwnnn';
+			$barChar['G'] = 'nnnnnwwnw';
+			$barChar['H'] = 'wnnnnwwnn';
+			$barChar['I'] = 'nnwnnwwnn';
+			$barChar['J'] = 'nnnnwwwnn';
+			$barChar['K'] = 'wnnnnnnww';
+			$barChar['L'] = 'nnwnnnnww';
+			$barChar['M'] = 'wnwnnnnwn';
+			$barChar['N'] = 'nnnnwnnww';
+			$barChar['O'] = 'wnnnwnnwn'; 
+			$barChar['P'] = 'nnwnwnnwn';
+			$barChar['Q'] = 'nnnnnnwww';
+			$barChar['R'] = 'wnnnnnwwn';
+			$barChar['S'] = 'nnwnnnwwn';
+			$barChar['T'] = 'nnnnwnwwn';
+			$barChar['U'] = 'wwnnnnnnw';
+			$barChar['V'] = 'nwwnnnnnw';
+			$barChar['W'] = 'wwwnnnnnn';
+			$barChar['X'] = 'nwnnwnnnw';
+			$barChar['Y'] = 'wwnnwnnnn';
+			$barChar['Z'] = 'nwwnwnnnn';
+			$barChar['-'] = 'nwnnnnwnw';
+			$barChar['.'] = 'wwnnnnwnn';
+			$barChar[' '] = 'nwwnnnwnn';
+			$barChar['*'] = 'nwnnwnwnn';
+			$barChar['$'] = 'nwnwnwnnn';
+			$barChar['/'] = 'nwnwnnnwn';
+			$barChar['+'] = 'nwnnnwnwn';
+			$barChar['%'] = 'nnnwnwnwn';
+		
+		//	$this->SetFont('Arial','',10);
+		//	$this->Text($xpos, $ypos + $height + 4, $code);
+		//	$this->SetFillColor(0);
+		
+			$code = '*'.strtoupper($code).'*';
+			for($i=0; $i<strlen($code); $i++){
+				$char = $code[$i];
+				if(!isset($barChar[$char])){
+					$this->Error('Invalid character in barcode: '.$char);
+				}
+				$seq = $barChar[$char];
+				for($bar=0; $bar<9; $bar++){
+					if($seq[$bar] == 'n'){
+						$lineWidth = $narrow;
+					}else{
+						$lineWidth = $wide;
+					}
+					if($bar % 2 == 0){
+						$this->Rect($xpos, $ypos, $lineWidth, $height, 'F');
+					}
+					$xpos += $lineWidth;
+				}
+				$xpos += $gap;
+			}
+		}
+
+}		
 
 	$display_heading = array('name'=> 'Product', 'ingredient'=> 'Ingredient','concentration'=> 'Concentration',);	
 	$formula_q = mysqli_query($conn, "SELECT * FROM formulas WHERE fid = '$fid' ORDER BY ingredient ASC");
@@ -57,6 +141,9 @@ function genBatchPDF($fid, $batchID, $bottle, $new_conc, $mg, $ver, $uploads_pat
 	$pdf->SetFont('Arial','B',10);
 	$pdf->MultiCell(250,10,$coverText);
 	
+	//Barcode
+	//$pdf->Code39(80,40,$batchID,1,10);
+
 	//Formula table
 	$pdf->AddPage();
 	$pdf->AliasNbPages();